+++ opentelemetry-sdk/src/metrics/internal/aggregate.rs
@@ -6,7 +6,7 @@ use std::{
     time::SystemTime,
 };
 
-use crate::metrics::{data::Aggregation, Temporality};
+use crate::metrics::{data::AggregatedMetrics, Temporality};
 use opentelemetry::time::now;
 use opentelemetry::KeyValue;
 
@@ -38,7 +38,7 @@ pub(crate) trait ComputeAggregation: Send + Sync + 'static {
     /// If no initial aggregation exists, `dest` will be `None`, in which case the
     /// returned option is expected to contain a new aggregation with the data from
     /// the current collection cycle.
-    fn call(&self, dest: Option<&mut dyn Aggregation>) -> (usize, Option<Box<dyn Aggregation>>);
+    fn call(&self, dest: Option<&mut AggregatedMetrics>) -> (usize, Option<AggregatedMetrics>);
 }
 
 /// Separate `measure` and `collect` functions for an aggregate.
@@ -205,7 +205,7 @@ impl<T: Number> AggregateBuilder<T> {
 mod tests {
     use crate::metrics::data::{
         ExponentialBucket, ExponentialHistogram, ExponentialHistogramDataPoint, Gauge,
-        GaugeDataPoint, Histogram, HistogramDataPoint, Sum, SumDataPoint,
+        GaugeDataPoint, Histogram, HistogramDataPoint, MetricData, Sum, SumDataPoint,
     };
     use std::vec;
 
@@ -215,7 +215,7 @@ mod tests {
     fn last_value_aggregation() {
         let AggregateFns { measure, collect } =
             AggregateBuilder::<u64>::new(Temporality::Cumulative, None).last_value(None);
-        let mut a = Gauge {
+        let mut a = MetricData::Gauge(Gauge {
             data_points: vec![GaugeDataPoint {
                 attributes: vec![KeyValue::new("a", 1)],
                 value: 1u64,
@@ -223,11 +223,15 @@ mod tests {
             }],
             start_time: Some(now()),
             time: now(),
-        };
+        })
+        .into();
         let new_attributes = [KeyValue::new("b", 2)];
         measure.call(2, &new_attributes[..]);
 
         let (count, new_agg) = collect.call(Some(&mut a));
+        let AggregatedMetrics::U64(MetricData::Gauge(a)) = a else {
+            unreachable!()
+        };
 
         assert_eq!(count, 1);
         assert!(new_agg.is_none());
@@ -241,7 +245,7 @@ mod tests {
         for temporality in [Temporality::Delta, Temporality::Cumulative] {
             let AggregateFns { measure, collect } =
                 AggregateBuilder::<u64>::new(temporality, None).precomputed_sum(true);
-            let mut a = Sum {
+            let mut a = MetricData::Sum(Sum {
                 data_points: vec![
                     SumDataPoint {
                         attributes: vec![KeyValue::new("a1", 1)],
@@ -262,11 +266,15 @@ mod tests {
                     Temporality::Delta
                 },
                 is_monotonic: false,
-            };
+            })
+            .into();
             let new_attributes = [KeyValue::new("b", 2)];
             measure.call(3, &new_attributes[..]);
 
             let (count, new_agg) = collect.call(Some(&mut a));
+            let AggregatedMetrics::U64(MetricData::Sum(a)) = a else {
+                unreachable!()
+            };
 
             assert_eq!(count, 1);
             assert!(new_agg.is_none());
@@ -283,7 +291,7 @@ mod tests {
         for temporality in [Temporality::Delta, Temporality::Cumulative] {
             let AggregateFns { measure, collect } =
                 AggregateBuilder::<u64>::new(temporality, None).sum(true);
-            let mut a = Sum {
+            let mut a = MetricData::Sum(Sum {
                 data_points: vec![
                     SumDataPoint {
                         attributes: vec![KeyValue::new("a1", 1)],
@@ -304,11 +312,15 @@ mod tests {
                     Temporality::Delta
                 },
                 is_monotonic: false,
-            };
+            })
+            .into();
             let new_attributes = [KeyValue::new("b", 2)];
             measure.call(3, &new_attributes[..]);
 
             let (count, new_agg) = collect.call(Some(&mut a));
+            let AggregatedMetrics::U64(MetricData::Sum(a)) = a else {
+                unreachable!()
+            };
 
             assert_eq!(count, 1);
             assert!(new_agg.is_none());
@@ -325,7 +337,7 @@ mod tests {
         for temporality in [Temporality::Delta, Temporality::Cumulative] {
             let AggregateFns { measure, collect } = AggregateBuilder::<u64>::new(temporality, None)
                 .explicit_bucket_histogram(vec![1.0], true, true);
-            let mut a = Histogram {
+            let mut a = MetricData::Histogram(Histogram {
                 data_points: vec![HistogramDataPoint {
                     attributes: vec![KeyValue::new("a1", 1)],
                     count: 2,
@@ -343,11 +355,15 @@ mod tests {
                 } else {
                     Temporality::Delta
                 },
-            };
+            })
+            .into();
             let new_attributes = [KeyValue::new("b", 2)];
             measure.call(3, &new_attributes[..]);
 
             let (count, new_agg) = collect.call(Some(&mut a));
+            let AggregatedMetrics::U64(MetricData::Histogram(a)) = a else {
+                unreachable!()
+            };
 
             assert_eq!(count, 1);
             assert!(new_agg.is_none());
@@ -368,7 +384,7 @@ mod tests {
         for temporality in [Temporality::Delta, Temporality::Cumulative] {
             let AggregateFns { measure, collect } = AggregateBuilder::<u64>::new(temporality, None)
                 .exponential_bucket_histogram(4, 20, true, true);
-            let mut a = ExponentialHistogram {
+            let mut a = MetricData::ExponentialHistogram(ExponentialHistogram {
                 data_points: vec![ExponentialHistogramDataPoint {
                     attributes: vec![KeyValue::new("a1", 1)],
                     count: 2,
@@ -395,11 +411,15 @@ mod tests {
                 } else {
                     Temporality::Delta
                 },
-            };
+            })
+            .into();
             let new_attributes = [KeyValue::new("b", 2)];
             measure.call(3, &new_attributes[..]);
 
             let (count, new_agg) = collect.call(Some(&mut a));
+            let AggregatedMetrics::U64(MetricData::ExponentialHistogram(a)) = a else {
+                unreachable!()
+            };
 
             assert_eq!(count, 1);
             assert!(new_agg.is_none());