+++ opentelemetry/src/lib.rs
@@ -246,7 +246,7 @@ pub mod global;
 
 pub mod baggage;
 
-mod context;
+pub mod context;
 
 pub use context::{Context, ContextGuard};
 
@@ -298,4 +298,10 @@ pub mod time {
     pub fn now() -> SystemTime {
         SystemTime::UNIX_EPOCH + std::time::Duration::from_millis(js_sys::Date::now() as u64)
     }
+}
+
+#[doc(hidden)]
+pub mod _private {
+    #[cfg(feature = "internal-logs")]
+    pub use tracing::{debug, error, info, warn}; // re-export
 }